requirements.txt
# ── Orchestration ─────────────────────────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0

# ── LLM / Embeddings (local via Ollama) ───────────────────────────────────────
ollama>=0.3.0

# ── Embeddings & Re-ranker ────────────────────────────────────────────────────
FlagEmbedding>=1.2.0           # BGE-M3 + BGE-Reranker-v2-m3
sentence-transformers>=3.0.0

# ── Vector Database ───────────────────────────────────────────────────────────
qdrant-client>=1.9.0

# ── Semantic Cache ────────────────────────────────────────────────────────────
redis>=5.0.0
numpy>=1.26.0

# ── Conversation Memory ───────────────────────────────────────────────────────

# ── Speech-to-Text ────────────────────────────────────────────────────────────
openai-whisper>=20231117
soundfile>=0.12.0
librosa>=0.10.0

# ── Voice Activity Detection ──────────────────────────────────────────────────
silero-vad>=5.1.0
torch>=2.0.0
torchaudio>=2.0.0

# ── Text-to-Speech ────────────────────────────────────────────────────────────
kokoro>=0.9.0                  # English / multilingual
# viettts                      # Vietnamese (install from GitHub)
requests>=2.31.0               # Zalo TTS API

# ── Document Ingestion ────────────────────────────────────────────────────────
pypdf>=4.0.0
python-docx>=1.1.0
unstructured>=0.14.0
markdown>=3.6

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0                # retry / back-off
structlog>=24.0.0              # structured logging
httpx>=0.27.0

# ── Visualization & Notebook Support ───────────────────────────────────────────
ipython>=8.0.0                 # For graph visualization in notebooks
